+++ src/utils/print.c
@@ -6,7 +6,7 @@
 /*   By: lraffin <lraffin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/16 01:33:48 by lraffin           #+#    #+#             */
-/*   Updated: 2021/12/19 01:44:00 by lraffin          ###   ########.fr       */
+/*   Updated: 2021/12/19 02:11:54 by lraffin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,13 +15,13 @@
 void	update_status(char *msg, t_philo *philo)
 {
 	t_data	*data;
-	long	time;
+	int	time;
 
 	data = philo->data;
 	time = get_time() - data->time.start;
 	pthread_mutex_lock(&data->write.mutex);
 	if (data->write.value == TRUE)
-		printf(" %-6lu ms | %4d | %s\n", time, philo->id + 1, msg);
+		printf(" %-6d ms | %4d | %s\n", time, philo->id + 1, msg);
 	pthread_mutex_unlock(&data->write.mutex);
 }
